Window Added
I added a window, but whenever I would try and run the code to open the window it would close immediately, So I have to deal with that too. Then I was able to add window rescale functionality. With that done, I went though and commented to make sure that I truly understood what was going on. I am like 80% sure that the comments are correct.