Vertex Data
Now the message might be a bit misleading, but its also technically true. I 2 structs that really only hold one triangle. I want to do the math myself, so I'm going to leave it like this. I did however have to write a vertex shader in glsl, which i dont understand, but it will come to me at some point.