Shader work
This description will be a bit less, and not really explain what I did, partially because I don't have time, but mainly because I really don't understand what I did. My guess is that after I set up the shaders at the top of the program, I had to actually compile it as a shader, so I made a shader program and attached the two shaders to it, now I have to figure out the rest of drawing this stupid triangle.